+++ src/App.js
@@ -4,7 +4,6 @@ import Header from "./components/shared/Header";
 import Footer from "./components/shared/Footer";
 import Newsletter from "./components/shared/Newsletter";
 import Main from "./components/Main";
-import Header2 from "./components/shared/Header2";
 import AdminHeader from "./components/admin/pages/adminHeader";
 import SideBar from "./components/admin/pages/sidebar";
 import AdminMain from "./components/admin/adminMain"
@@ -14,7 +13,7 @@ class App extends Component {
     constructor(props){
         super(props);
         this.state = {
-            isAdmin : true,
+            isAdmin : false,
         }
     }
 
@@ -30,7 +29,7 @@ class App extends Component {
                             <AdminMain/>
                         </div> :
                         <div>
-                            <Header2/>
+                            <Header/>
                             <Main/>
                             <Newsletter/>
                         </div>
